Improves GAME's scoring module, mostly from the following two perspectives: 1. Separate the logic of scoring and model evaluation 2. Created a more user-friendly output of the scoring job. Also improved test coverage for scoring related components.
