+++ photon-ml/src/integTest/resources/GameIntegTest/gameModel/fixed-effect/globalShard/id-info
@@ -1,0 +1,1 @@
+globalShard
