+++ photon-ml/src/integTest/resources/GameIntegTest/gameModel/random-effect/songId-songShard/id-info
@@ -1,0 +1,2 @@
+songId
+songShard
