+++ photon-ml/src/integTest/resources/GameIntegTest/gameModel/random-effect/userId-userShard/id-info
@@ -1,0 +1,2 @@
+userId
+userShard
