+++ photon-ml/src/integTest/resources/GameIntegTest/input/feature-lists/songFeatures
@@ -1,0 +1,20 @@
+s	18
+s	5
+s	12
+s	0
+s	4
+s	8
+s	15
+s	11
+s	3
+s	19
+s	2
+s	14
+s	16
+s	9
+s	10
+s	6
+s	17
+s	13
+s	7
+s	1
