+++ photon-ml/src/integTest/resources/GameIntegTest/input/feature-lists/userFeatures
@@ -1,0 +1,20 @@
+u	6
+u	13
+u	19
+u	1
+u	18
+u	9
+u	10
+u	5
+u	16
+u	2
+u	12
+u	4
+u	15
+u	3
+u	11
+u	0
+u	14
+u	17
+u	8
+u	7
