+++ photon-ml/src/integTest/scala/com/linkedin/photon/ml/avro/DataProcessingUtilsTest.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2016 LinkedIn Corp. All rights reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package com.linkedin.photon.ml.avro
+
+import org.apache.avro.generic.GenericRecord
+import org.apache.spark.rdd.RDD
+
+import com.linkedin.photon.ml.util.Utils
+
+object DataProcessingUtilsTest {
+
+  def loadUidsWithResponsesFromGenericRecords(records: RDD[GenericRecord]): RDD[(String, Double)] = {
+    records.map { record =>
+      val uid = Utils.getStringAvro(record, AvroFieldNames.UID)
+      val response = Utils.getDoubleAvro(record, AvroFieldNames.RESPONSE)
+      (uid, response)
+    }
+  }
+}
