+++ photon-ml/src/main/scala/com/linkedin/photon/ml/cli/game/scoring/Driver.scala
@@ -14,6 +14,7 @@
  */
 package com.linkedin.photon.ml.cli.game.scoring
 
+
 import scala.collection.Map
 
 import org.apache.hadoop.fs.Path
@@ -21,20 +22,17 @@ import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 
 import com.linkedin.photon.ml.avro.AvroUtils
-import com.linkedin.photon.ml.avro.data.{DataProcessingUtils, NameAndTermFeatureSetContainer, NameAndTerm}
-import com.linkedin.photon.ml.avro.model.ModelProcessingUtils
+import com.linkedin.photon.ml.avro.data.{DataProcessingUtils, NameAndTerm, NameAndTermFeatureSetContainer}
+import com.linkedin.photon.ml.avro.data.ScoreProcessingUtils
 import com.linkedin.photon.ml.constants.StorageLevel
-import com.linkedin.photon.ml.data.GameDatum
-import com.linkedin.photon.ml.evaluation.{RMSEEvaluator, BinaryClassificationEvaluator}
+import com.linkedin.photon.ml.data.{GameDatum, KeyValueScore}
 import com.linkedin.photon.ml.SparkContextConfiguration
-import com.linkedin.photon.ml.supervised.TaskType._
+import com.linkedin.photon.ml.avro.model.ModelProcessingUtils
 import com.linkedin.photon.ml.util._
 
 
 /**
  * Driver for GAME full model scoring
- *
- * @author xazhang
  */
 class Driver(val params: Params, val sparkContext: SparkContext, val logger: PhotonLogger) {
 
@@ -70,110 +68,88 @@ class Driver(val params: Params, val sparkContext: SparkContext, val logger: Pho
   }
 
   /**
-   * Builds a GAME dataset according to input data configuration
+   * Builds a GAME data set according to input data configuration
    *
    * @param featureShardIdToFeatureMapMap a map of shard id to feature map
-   * @return the prepared GAME dataset
+   * @return the prepared GAME data set
    */
   protected def prepareGameDataSet(featureShardIdToFeatureMapMap: Map[String, Map[NameAndTerm, Int]])
-  : RDD[(Long, GameDatum)] = {
+  : (RDD[(Long, String)], RDD[(Long, GameDatum)]) = {
 
-    val recordsPath = dateRangeOpt match {
-      case Some(dateRange) =>
-        val range = DateRange.fromDates(dateRange)
-        IOUtils.getInputPathsWithinDateRange(inputDirs, range, hadoopConfiguration, errorOnMissing = false)
-      case None => inputDirs.toSeq
+    val recordsPath = (dateRangeOpt, dateRangeDaysAgoOpt) match {
+      // Specified as date range
+      case (Some(trainDateRange), None) =>
+        val dateRange = DateRange.fromDates(trainDateRange)
+        IOUtils.getInputPathsWithinDateRange(inputDirs, dateRange, hadoopConfiguration, errorOnMissing = false)
+
+      // Specified as a range of start days ago - end days ago
+      case (None, Some(trainDateRangeDaysAgo)) =>
+        val dateRange = DateRange.fromDaysAgo(trainDateRangeDaysAgo)
+        IOUtils.getInputPathsWithinDateRange(inputDirs, dateRange, hadoopConfiguration, errorOnMissing = false)
+
+      // Both types specified: illegal
+      case (Some(_), Some(_)) =>
+        throw new IllegalArgumentException(
+          "Both trainDateRangeOpt and trainDateRangeDaysAgoOpt given. You must specify date ranges using only one " +
+              "format.")
+
+      // No range specified, just use the train dir
+      case (None, None) => inputDirs.toSeq
     }
-    logger.debug(s"Avro records paths:\n${recordsPath.mkString("\n")}")
+    logger.debug(s"Input records paths:\n${recordsPath.mkString("\n")}")
     val records = AvroUtils.readAvroFiles(sparkContext, recordsPath, parallelism)
+    val recordsWithUniqueId = records.zipWithUniqueId().map(_.swap)
     val globalDataPartitioner = new LongHashPartitioner(records.partitions.length)
 
-    val gameDataSet = DataProcessingUtils.getGameDataSetFromGenericRecords(records,
-      featureShardIdToFeatureSectionKeysMap, featureShardIdToFeatureMapMap, randomEffectIdSet)
-        .partitionBy(globalDataPartitioner)
-        .setName("Scoring Game data set")
-        .persist(StorageLevel.FREQUENT_REUSE_RDD_STORAGE_LEVEL)
+    val gameDataSetWithUIDs = DataProcessingUtils.getGameDataSetWithUIDFromGenericRecords(
+      recordsWithUniqueId,
+      featureShardIdToFeatureSectionKeysMap,
+      featureShardIdToFeatureMapMap,
+      randomEffectIdSet,
+      isResponseRequired = false)
+      .partitionBy(globalDataPartitioner)
+      .setName("Game data set with UIDs for scoring")
+      .persist(StorageLevel.INFREQUENT_REUSE_RDD_STORAGE_LEVEL)
+
+    val gameDataSet = gameDataSetWithUIDs.mapValues(_._1)
+    val uids = gameDataSetWithUIDs.mapValues(_._2)
 
     // Log some simple summary info on the Game data set
-    logger.debug(s"Summary for the validating Game data set")
-    val numSamples = gameDataSet.count()
+    logger.debug(s"Summary for the GAME data set")
+    val numSamples = gameDataSetWithUIDs.count()
     logger.debug(s"numSamples: $numSamples")
-    val responseSum = gameDataSet.values.map(_.response).sum()
-    logger.debug(s"responseSum: $responseSum")
-    val weightSum = gameDataSet.values.map(_.weight).sum()
-    logger.debug(s"weightSum: $weightSum")
-    val randomEffectIdToIndividualIdMap = gameDataSet.values.first().randomEffectIdToIndividualIdMap
-    randomEffectIdToIndividualIdMap.keySet.foreach { randomEffectId =>
-      val dataStats = gameDataSet.values.map { gameData =>
+    randomEffectIdSet.foreach { randomEffectId =>
+      val numSamplesStats = gameDataSet.map { case (_, gameData) =>
         val individualId = gameData.randomEffectIdToIndividualIdMap(randomEffectId)
-        (individualId, (gameData.response, 1))
-      }.reduceByKey { case ((responseSum1, numSample1), (responseSum2, numSample2)) =>
-        (responseSum1 + responseSum2, numSample1 + numSample2)
-      }.cache()
-      val responseSumStats = dataStats.values.map(_._1).stats()
-      val numSamplesStats = dataStats.values.map(_._2).stats()
-      logger.debug(s"numSamplesStats for $randomEffectId: $numSamplesStats")
-      logger.debug(s"responseSumStats for $randomEffectId: $responseSumStats")
+        (individualId, 1)
+      }
+        .reduceByKey(_ + _)
+        .values
+        .stats()
+      logger.debug(s"numSamples for $randomEffectId: $numSamplesStats")
     }
-
-    gameDataSet
+    (uids, gameDataSet)
   }
 
   /**
-   * Score and write results to HDFS
+   * Score the game data set with the game model
    * @param featureShardIdToFeatureMapMap a map of shard id to feature map
-   * @param gameDataSet the input dataset
+   * @param gameDataSet the game data set
    * @return the scores
    */
   //todo: make the number of files written to HDFS to be configurable
-  protected def scoreAndWriteScoreToHDFS(
+  protected def scoreGameDataSet(
       featureShardIdToFeatureMapMap: Map[String, Map[NameAndTerm, Int]],
-      gameDataSet: RDD[(Long, GameDatum)]): RDD[(Long, Double)] = {
+      gameDataSet: RDD[(Long, GameDatum)]): KeyValueScore = {
 
     val gameModel = ModelProcessingUtils.loadGameModelFromHDFS(featureShardIdToFeatureMapMap, gameModelInputDir,
       sparkContext)
 
     logger.debug(s"Loaded game model summary:\n${gameModel.toSummaryString}")
 
-    val scores = gameModel.score(gameDataSet).scores
-        .setName("Scores").persist(StorageLevel.FREQUENT_REUSE_RDD_STORAGE_LEVEL)
-    val scoredItems = scores.join(gameDataSet).map { case (_, (score, gameData)) =>
-      val ids = gameData.randomEffectIdToIndividualIdMap.values
-      val label = gameData.response
-      ScoredItem(ids, score, label)
-    }.setName("Scored item").persist(StorageLevel.FREQUENT_REUSE_RDD_STORAGE_LEVEL)
-
-    val numScoredItems = scoredItems.count()
-    val scoresDir = new Path(outputDir, Driver.SCORES).toString
-    // Should always materialize the scoredItems first (e.g., count()) before the coalesce happens
-    scoredItems.coalesce(numFiles).saveAsTextFile(scoresDir)
-    logger.debug(s"Number of scored items: $numScoredItems")
-
-    scores
-  }
-
-  /**
-   * Evaluate and log metrics for the GAME dataset
-   *
-   * @param gameDataSet the input dataset
-   * @param scores the scores
-   */
-  protected def evaluateAndLog(gameDataSet: RDD[(Long, GameDatum)], scores: RDD[(Long, Double)]) {
-
-    val validatingLabelAndOffsets = gameDataSet.mapValues(gameData => (gameData.response, gameData.offset))
-    val metric =  taskType match {
-      case LOGISTIC_REGRESSION =>
-        new BinaryClassificationEvaluator(validatingLabelAndOffsets).evaluate(scores)
-      case LINEAR_REGRESSION =>
-        val validatingLabelAndOffsetAndWeights = validatingLabelAndOffsets.mapValues { case (label, offset) =>
-          (label, offset, 1.0)
-        }
-        new RMSEEvaluator(validatingLabelAndOffsetAndWeights).evaluate(scores)
-      case _ =>
-        throw new UnsupportedOperationException(s"Task type: $taskType is not supported to create validating " +
-            s"evaluator")
-    }
-    logger.info(s"Evaluation metric: $metric")
+    val scores = gameModel.score(gameDataSet)
+    val offsets = new KeyValueScore(gameDataSet.mapValues(_.offset))
+    scores + offsets
   }
 
   /**
@@ -187,25 +163,37 @@ class Driver(val params: Params, val sparkContext: SparkContext, val logger: Pho
     logger.info(s"Time elapsed after preparing feature maps: $initializationTime (s)\n")
 
     startTime = System.nanoTime()
-    val gameDataSet = prepareGameDataSet(featureShardIdToFeatureMapMap)
+    val (uids, gameDataSet) = prepareGameDataSet(featureShardIdToFeatureMapMap)
     val gameDataSetPreparationTime = (System.nanoTime() - startTime) * 1e-9
     logger.info(s"Time elapsed after game data set preparation: $gameDataSetPreparationTime (s)\n")
 
     startTime = System.nanoTime()
-    val scores = scoreAndWriteScoreToHDFS(featureShardIdToFeatureMapMap, gameDataSet)
+    val scores = scoreGameDataSet(featureShardIdToFeatureMapMap, gameDataSet)
+    val scoredItems = uids.join(scores.scores).map { case (_, (uid, score)) => ScoredItem(uid, score) }
+    scoredItems.setName("Scored items").persist(StorageLevel.INFREQUENT_REUSE_RDD_STORAGE_LEVEL)
+    val numScoredItems = scoredItems.count()
+    logger.info(s"Number of scored items: $numScoredItems (s)\n")
+    val scoresDir = new Path(outputDir, Driver.SCORES).toString
+    val scoredItemsToBeSaved =
+      if (numOutputFilesForScores > 0 && numOutputFilesForScores != scoredItems.partitions.length) {
+        scoredItems.coalesce(numOutputFilesForScores)
+      } else {
+        scoredItems
+      }
+
+    Utils.deleteHDFSDir(scoresDir, hadoopConfiguration)
+    ScoreProcessingUtils.saveScoredItemsToHDFS(scoredItemsToBeSaved, modelId = "", scoresDir)
     val scoringTime = (System.nanoTime() - startTime) * 1e-9
     logger.info(s"Time elapsed scoring and writing scores to HDFS: $scoringTime (s)\n")
 
-    startTime = System.nanoTime()
-    evaluateAndLog(gameDataSet, scores)
     val postprocessingTime = (System.nanoTime() - startTime) * 1e-9
     logger.info(s"Time elapsed after evaluation: $postprocessingTime (s)\n")
   }
 }
 
 object Driver {
-  private val SCORES = "scores"
-  private val LOGS = "logs"
+  protected[scoring] val SCORES = "scores"
+  protected[scoring] val LOGS = "logs"
 
   /**
    * Main entry point
