+++ photon-test/src/main/scala/com/linkedin/photon/ml/test/CommonTestUtils.scala
@@ -49,13 +49,13 @@ object CommonTestUtils {
         if (x.size != y.size) {
           false
         } else {
-          x.filter {
+          x.count {
             case (id, bounds) =>
               y.get(id) match {
                 case Some(w: (Double, Double)) => w == bounds
                 case None => false
               }
-          }.size == x.size
+          } == x.size
         }
       case (None, None) => true
       case (None, _) => false
